refactor: unify firebase initialization
